consumer_four/requirements.txt
Flask==2.0.2
pika==1.2.0
mysql-connector-python==8.0.25